restructured init into folder
